requirements.txt
docopt
gitpython
nose
PyGithub
scikit-learn
wheel
